student_teacher_app/requirements.txt
packages = [
django,
psycopg2,
djangorestframework,
]